refactor(module): rename modules to module
