+++ src/debmagic/_module/autodetect.py
