+++ src/debmagic/_module/autotools.py
