+++ src/debmagic/_module/default.py
